primera subida 4 febrero

+++ ft_printmaxhex.c
@@ -6,17 +6,17 @@
 /*   By: ismonter <ismonter@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/03 20:32:03 by ismonter          #+#    #+#             */
-/*   Updated: 2026/02/03 20:33:08 by ismonter         ###   ########.fr       */
+/*   Updated: 2026/02/04 11:30:30 by ismonter         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 
-int	ft_printmaxhex(unsigned int nbr)
+int	ft_printmaxhex(unsigned long nbr)
 {
-	int	n;
-	const char *dic_hex = "0123456789ABCDEF";
-	char a;
+	int			n;
+	const char	*dic_hex = "0123456789ABCDEF";
+	char		a;
 
 	n = 0;
 	if (nbr >= 16)
